requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.6
GeoAlchemy2==0.16.0
geojson==3.1.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
numpy==2.1.3
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
pytest-asyncio==0.24.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
